Removed the original completion contributors and keeping only the reworked and relevant ones
