+++ intelij-plugin/src/main/resources/META-INF/plugin.xml
@@ -43,14 +43,7 @@
 
         <!-- Completion -->
         <completion.contributor language="yang-pantheon"
-          implementationClass="tech.pantheon.yanginator.plugin.completionBeta.YangCompletionContributor"/>
-<!--
-        <completion.contributor language="yang-pantheon"
-                                implementationClass="tech.pantheon.yanginator.plugin.completion.YangCompletionContributor"/>
-        <lang.foldingBuilder language="yang-pantheon"
-                             implementationClass="tech.pantheon.yanginator.plugin.completion.YangCompletionContributorBuilder"/>
-        <psi.treeChangeListener implementation="tech.pantheon.yanginator.plugin.completion.YangPsiTreeChangeListener"/>
--->
+          implementationClass="tech.pantheon.yanginator.plugin.completion.YangCompletionContributor"/>
 
         <!-- Reference -->
         <psi.referenceContributor language="yang-pantheon"
